+++ integration_test/src/v17/wallet.rs
@@ -83,7 +83,7 @@ macro_rules! impl_test_v17__getbalance {
     () => {
         #[test]
         fn get_balance() {
-            use client::json::model;
+            use client::types::model;
 
             let bitcoind = $crate::bitcoind_with_default_wallet();
             let json = bitcoind.client.get_balance().expect("getbalance");
@@ -125,7 +125,7 @@ macro_rules! impl_test_v17__gettransaction {
         #[test]
         fn get_transaction() {
             use bitcoin::Amount;
-            use client::json::model;
+            use client::types::model;
 
             let bitcoind = $crate::bitcoind_with_default_wallet();
             let address = bitcoind.client.new_address().expect("failed to create new address");