+++ client/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.4.0 - 2024-11-14
+
+- Add support for Bitcoin Core v28
+- Re-name the repository from `rust-bitcoind-josn-rpc` to `corepc`.
+- Re-name the crate from `bitcoind-josn-rpc-client` to `corepc-client`.
+
 # 0.3.0 - 2024-06-21
 
 - Fix bugs in `AddressType`